requirements/production.txt
# Pro-tip: Try not to put anything here. Avoid dependencies in
#	production that aren't in development.
-r base.txt



# WSGI Handler
# ------------------------------------------------
gevent==1.2.2
gunicorn==19.7.1

# Static and Media Storage
# ------------------------------------------------
boto3==1.4.7
django-storages==1.6.5


# Email backends for Mailgun, Postmark, SendGrid and more
# -------------------------------------------------------
django-anymail==1.0

# Raven is the Sentry client
# --------------------------
raven==6.1.0

# Opbeat agent for performance monitoring
# -----------------------------------------
opbeat==3.5.2


-e git+git://github.com/handycodejob/django-coin-hive.git

